requirements.txt
# Google Authentication & API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0

# HTML Parsing
beautifulsoup4>=4.12.0

# lxml parser for BeautifulSoup (recommended)
lxml>=4.9.0
